requirements.txt
certifi==2018.4.16
Cython==0.28.4
dj-database-url==0.4.2
Django==2.0.5
djangorestframework==3.8.2
gunicorn==19.9.0
mkl-fft==1.0.4
mkl-random==1.0.1
numpy==1.15.0
pandas==0.23.3
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
wincertstore==0.2
